Fix Fetch API for random pokemon search
I changed the original idea of creating an array of urls by fetching just the pokemon names and urls and then looping over the elements of the array to get each pokemon and its stats. Now I'm directly request at the /pokemon end point and creating an array of pokemon Objects to be displayed.